backend/requirements.txt
Django==3.2.16
djangorestframework==3.14.0
djoser==2.2.2
psycopg2-binary==2.9.3
Pillow==9.0.0
django-filter==23.3